made more analyzer exception to include inner exceptions
